Added memory offset to PREM tasks, need to remove print

+++ src/hypervisor.c
@@ -1,8 +1,8 @@
 #include <hypervisor.h>
 
-int hypercall(enum hypervisor_actions action, int arg0, int arg1, int arg2) {
-    int hypercall_id = HYPERCALL_BASE_VALUE | action;
-    int ret;
+uint64_t hypercall(enum hypervisor_actions action, int arg0, int arg1, int arg2) {
+    uint64_t hypercall_id = HYPERCALL_BASE_VALUE | action;
+    uint64_t ret;
 
     __asm__ volatile (
         "hvc #0\n\t"
