+++ src/inc/hypervisor.h
@@ -1,6 +1,8 @@
 #ifndef __HYPERVISOR_H__
 #define __HYPERVISOR_H__
 
+#include <FreeRTOS.h>
+
 /* Hypervisor base value (aarch64) */
 #define HYPERCALL_BASE_VALUE 0x86000000 | 0x40000000
 
@@ -8,7 +10,7 @@
 enum hypervisor_actions { HC_IPC = 1, HC_REQUEST_MEM_ACCESS = 2, HC_REVOKE_MEM_ACCESS = 3, HC_GET_CPU_ID = 4, HC_NOTIFY_CPU = 5, HC_EMPTY_CALL = 6, HC_REQUEST_MEM_ACCESS_TIMER = 7 };
 
 /* Hypercall with the specified action and arguments */
-int hypercall(enum hypervisor_actions, int arg0, int arg1, int arg2);
+uint64_t hypercall(enum hypervisor_actions, int arg0, int arg1, int arg2);
 
 /* Macros for memory request and revoke */
 #define request_memory_access(prio) hypercall(HC_REQUEST_MEM_ACCESS, prio, 0, 0)
